Add event module and their components, Re-organize directories
