+++ src/app/_components/footer.tsx
@@ -1,4 +1,4 @@
-import { WebComponentProps } from '../../../types';
+import { WebComponentProps } from '../../types';
 
 export default function Footer(props: WebComponentProps) {
 	const { currentPage } = props;
