+++ src/app/_components/module/event/searchBox.tsx
@@ -8,8 +8,11 @@ interface MultiFunctionButtonProps {
 
 function MultiFunctionButton(props: MultiFunctionButtonProps) {
 	const { children, width, icon } = props;
-	const buttonClasses = `w-[${width}px] bg-red-200 h-full rounded-[33px]`;
-	return <div className={buttonClasses}>{children}</div>;
+	return (
+		<div className="bg-red-200 h-full rounded-[33px]" style={{ width: width }}>
+			{children}
+		</div>
+	);
 }
 
 export default function SearchBox() {
@@ -18,7 +21,7 @@ export default function SearchBox() {
 			className="flex flex-row h-[164px] w-full justify-between"
 			initial={{ opacity: 0 }}
 			animate={{ opacity: 1 }}
-			exit={{ opacity: 0 }}
+			exit={{ scaleY: 0 }}
 		>
 			<MultiFunctionButton width={477}>Create</MultiFunctionButton>
 			<MultiFunctionButton width={685}>Search</MultiFunctionButton>
