--- src/app/_components/title.tsx
@@ -1,9 +1,0 @@
-import SearchBox from './specific/searchBox';
-
-export default function Title() {
-	return (
-		<div className="pt-[372px]">
-			<SearchBox />
-		</div>
-	);
-}
