--- src/app/_components/utils/groupComponent.tsx
@@ -1,3 +1,0 @@
-export default function GroupComponent(props: GroupComponentProps) {
-	return <>{props.children}</>;
-}
