+++ src/app/page.tsx
@@ -1,15 +1,21 @@
 'use client';
 
-import Navbar from './_components/web/navbar';
-import Head from './_components/web/head';
-import Body from './_components/web/body';
-import Footer from './_components/web/footer';
+import Navbar from './_components/navbar';
+import Body from './_components/body';
+import Footer from './_components/footer';
 import { useEffect, useState } from 'react';
-import EventCardModal from './_components/modal/eventCardModal';
-import { motion, AnimatePresence } from 'framer-motion';
+import { EventCardProps } from '@/types';
+
+const sampleEvents: EventCardProps[] = [
+	{ id: 0, title: 'Event 1', description: 'Description 1' },
+	{ id: 1, title: 'Event 2', description: 'Description 2' },
+	{ id: 2, title: 'Event 3', description: 'Description 3' },
+	{ id: 3, title: 'Event 4', description: 'Description 4' },
+];
 
 export default function App() {
 	const [currentPage, setCurrentPage] = useState('');
+	const [events, setEvents] = useState<EventCardProps[]>(sampleEvents);
 
 	useEffect(() => {
 		console.log('Current Page: ', currentPage);
@@ -18,11 +24,10 @@ export default function App() {
 	return (
 		<div>
 			<Navbar currentPage={currentPage} setCurrentPage={setCurrentPage} />
-			<div className="flex flex-col justify-center items-center w-full mx-auto">
-				<Head currentPage={currentPage} setCurrentPage={setCurrentPage} />
-				<Body currentPage={currentPage} setCurrentPage={setCurrentPage} />
-				<Footer currentPage={currentPage} setCurrentPage={setCurrentPage} />
+			<div className="flex flex-col justify-center items-center w-full">
+				<Body currentPage={currentPage} setCurrentPage={setCurrentPage} events={events} setEvents={setEvents} />
 			</div>
+			<Footer currentPage={currentPage} setCurrentPage={setCurrentPage} />
 		</div>
 	);
 }
